Remove announcement module from sample manager distribution, #6335 (#646)
Move email preference handling and schema to core.

Drop PageTypes and EmailFormats tables -- unused
