+++ announcements/resources/schemas/comm.xml
@@ -458,40 +458,6 @@
       </column>
     </columns>
   </table>
-  <table tableName="EmailFormats" tableDbType="TABLE">
-    <description>Contains one row per email format option (HTML/plain text)</description>
-    <columns>
-      <column columnName="EmailFormatId"/>
-      <column columnName="EmailFormat"/>
-    </columns>
-  </table>
-  <table tableName="EmailOptions" tableDbType="TABLE">
-    <description>Contains one row per subscription type, such as all messages in a given forum, no email at all, etc and daily digest options</description>
-      <columns>
-      <column columnName="EmailOptionId"/>
-      <column columnName="EmailOption"/>
-      <column columnName="Type"/>
-    </columns>
-  </table>
-  <table tableName="PageTypes" tableDbType="TABLE">
-    <columns>
-      <column columnName="PageTypeId"/>
-      <column columnName="PageType"/>
-    </columns>
-  </table>
-  <table tableName="EmailPrefs" tableDbType="TABLE">
-    <description>Contains one row per user-specific subscription</description>
-    <columns>
-      <column columnName="Container"/>
-      <column columnName="UserId"/>
-      <column columnName="Type"/>
-      <column columnName="EmailOptionId"/>
-      <column columnName="EmailFormatId"/>
-      <column columnName="PageTypeId"/>
-      <column columnName="LastModifiedBy"/>
-      <column columnName="SrcIdentifier"/>
-    </columns>
-  </table>
   <table tableName="UserList" tableDbType="TABLE">
       <description>Contains one row per user-specific thread-level subscription.</description>
     <columns>
