+++ announcements/src/org/labkey/announcements/NotificationSettingsImporterFactory.java
@@ -18,8 +18,8 @@ package org.labkey.announcements;
 import org.jetbrains.annotations.NotNull;
 import org.labkey.announcements.model.AnnouncementManager;
 import org.labkey.api.admin.AbstractFolderImportFactory;
-import org.labkey.api.admin.FolderImporter;
 import org.labkey.api.admin.FolderArchiveDataTypes;
+import org.labkey.api.admin.FolderImporter;
 import org.labkey.api.admin.ImportContext;
 import org.labkey.api.admin.ImportException;
 import org.labkey.api.files.FileContentDefaultEmailPref;
@@ -28,7 +28,7 @@ import org.labkey.api.notification.EmailService;
 import org.labkey.api.pipeline.PipelineJob;
 import org.labkey.api.pipeline.PipelineJobWarning;
 import org.labkey.api.writer.VirtualFile;
-import org.labkey.folder.xml.FolderDocument;
+import org.labkey.folder.xml.FolderDocument.Folder;
 import org.labkey.folder.xml.NotificationsType;
 
 import java.util.Collection;
@@ -46,7 +46,7 @@ public class NotificationSettingsImporterFactory extends AbstractFolderImportFac
         return new NotificationSettingsImporter();
     }
 
-    public class NotificationSettingsImporter implements  FolderImporter<FolderDocument.Folder>
+    public static class NotificationSettingsImporter implements FolderImporter<Folder>
     {
         @Override
         public String getDataType()
@@ -61,7 +61,7 @@ public class NotificationSettingsImporterFactory extends AbstractFolderImportFac
         }
 
         @Override
-        public void process(PipelineJob job, ImportContext<FolderDocument.Folder> ctx, VirtualFile root) throws Exception
+        public void process(PipelineJob job, ImportContext<Folder> ctx, VirtualFile root) throws Exception
         {
             if (isValidForImportArchive(ctx))
             {
@@ -94,13 +94,13 @@ public class NotificationSettingsImporterFactory extends AbstractFolderImportFac
 
         @NotNull
         @Override
-        public Collection<PipelineJobWarning> postProcess(ImportContext<FolderDocument.Folder> ctx, VirtualFile root)
+        public Collection<PipelineJobWarning> postProcess(ImportContext<Folder> ctx, VirtualFile root)
         {
             return Collections.emptyList();
         }
 
         @Override
-        public boolean isValidForImportArchive(ImportContext<FolderDocument.Folder> ctx) throws ImportException
+        public boolean isValidForImportArchive(ImportContext<Folder> ctx) throws ImportException
         {
             return ctx.getXml() != null && ctx.getXml().isSetNotifications();
         }
