+++ announcements/src/org/labkey/announcements/config/AnnouncementEmailConfig.java
@@ -39,10 +39,17 @@ public class AnnouncementEmailConfig extends AbstractConfigTypeProvider
         return TYPE;
     }
 
+    @Override
     public String getName()
     {
         // appears in the config tab
         return getType();
+    }
+
+    @Override
+    public String getSettingsColumnName()
+    {
+        return "MessageSettings";
     }
 
     @Override
