+++ announcements/src/org/labkey/announcements/model/AnnouncementDigestProvider.java
@@ -26,7 +26,6 @@ import org.labkey.api.data.SQLFragment;
 import org.labkey.api.data.SqlSelector;
 import org.labkey.api.message.digest.MessageDigest;
 import org.labkey.api.security.User;
-import org.labkey.api.security.UserManager;
 import org.labkey.api.util.ConfigurationException;
 import org.labkey.api.util.DateUtil;
 import org.labkey.api.util.MailHelper;
@@ -153,24 +152,27 @@ public class AnnouncementDigestProvider implements MessageDigest.Provider
             super(NAME, DEFAULT_SUBJECT, loadBody(), DEFAULT_DESCRIPTION, ContentType.HTML);
             setEditableScopes(EmailTemplate.Scope.SiteOrFolder);
 
-            _replacements.add(new ReplacementParam<String>("folderName", String.class, "Folder that user subscribed to", ContentType.Plain)
+            _replacements.add(new ReplacementParam<>("folderName", String.class, "Folder that user subscribed to", ContentType.Plain)
             {
+                @Override
                 public String getValue(Container c)
                 {
                     return c.getPath();
                 }
             });
 
-            _replacements.add(new ReplacementParam<String>("postList", String.class, "List of new posts", ContentType.HTML)
+            _replacements.add(new ReplacementParam<>("postList", String.class, "List of new posts", ContentType.HTML)
             {
+                @Override
                 public String getValue(Container c)
                 {
                     return posts;
                 }
             });
 
-            _replacements.add(new ReplacementParam<String>("reasonFooter", String.class, "Footer message explaining why user is receiving this digest", ContentType.HTML)
+            _replacements.add(new ReplacementParam<>("reasonFooter", String.class, "Footer message explaining why user is receiving this digest", ContentType.HTML)
             {
+                @Override
                 public String getValue(Container c)
                 {
                     return reasonForEmail;
@@ -180,6 +182,7 @@ public class AnnouncementDigestProvider implements MessageDigest.Provider
             _replacements.addAll(super.getValidReplacements());
         }
 
+        @Override
         public List<ReplacementParam> getValidReplacements()
         {
             return _replacements;
