+++ announcements/src/org/labkey/announcements/model/AnnouncementManager.java
@@ -49,6 +49,8 @@ import org.labkey.api.data.Table;
 import org.labkey.api.data.TableInfo;
 import org.labkey.api.data.TableSelector;
 import org.labkey.api.message.settings.MessageConfigService;
+import org.labkey.api.message.settings.MessageConfigService.ConfigTypeProvider;
+import org.labkey.api.message.settings.MessageConfigService.NotificationOption;
 import org.labkey.api.query.FieldKey;
 import org.labkey.api.search.SearchService;
 import org.labkey.api.security.SecurityManager;
@@ -185,8 +187,9 @@ public class AnnouncementManager
         }
     }
 
-    private static MessageConfigService.ConfigTypeProvider _configProvider;
-    public static MessageConfigService.ConfigTypeProvider getAnnouncementConfigProvider()
+    private static ConfigTypeProvider _configProvider;
+
+    public static ConfigTypeProvider getAnnouncementConfigProvider()
     {
         if (_configProvider == null)
         {
@@ -431,6 +434,7 @@ public class AnnouncementManager
                     }
                 }
 
+                //noinspection CallToThreadRun
                 emailer.run();  // We're already in a background thread... no need to start another one
             }
         });
@@ -562,7 +566,7 @@ public class AnnouncementManager
         return new TableSelector( _comm.getTableInfoAnnouncements(), SimpleFilter.createContainerFilter(c), null).getRowCount();
     }
 
-    public static MessageConfigService.NotificationOption[] getEmailOptions()
+    public static Collection<? extends NotificationOption> getEmailOptions()
     {
         return getAnnouncementConfigProvider().getOptions();
     }
@@ -643,7 +647,6 @@ public class AnnouncementManager
     public static void purgeContainer(Container c)
     {
         // Attachments are handled by AttachmentServiceImpl
-        ContainerUtil.purgeTable(_comm.getTableInfoEmailPrefs(), c, null);
         ContainerUtil.purgeTable(_comm.getTableInfoAnnouncements(), c, null);
     }
 
@@ -964,6 +967,7 @@ public class AnnouncementManager
 
             _replacements.add(new ReplacementParam<>("createdByUser", String.class, "User that generated the message", ContentType.Plain)
             {
+                @Override
                 public String getValue(Container c)
                 {
                     if (notificationBean == null)
@@ -975,6 +979,7 @@ public class AnnouncementManager
 
             _replacements.add(new ReplacementParam<>("createdOrResponded", String.class, "Created or Responded to a message", ContentType.HTML)
             {
+                @Override
                 public String getValue(Container c)
                 {
                     if (notificationBean == null)
@@ -985,6 +990,7 @@ public class AnnouncementManager
 
             _replacements.add(new ReplacementParam<>("messageDatetime", Date.class, "Date and time the message is created", ContentType.HTML)
             {
+                @Override
                 public Date getValue(Container c)
                 {
                     if (notificationBean == null)
@@ -995,6 +1001,7 @@ public class AnnouncementManager
 
             _replacements.add(new ReplacementParam<>("messageUrl", String.class, "Link to the original message", ContentType.Plain)
             {
+                @Override
                 public String getValue(Container c)
                 {
                     return messageUrl;
@@ -1003,6 +1010,7 @@ public class AnnouncementManager
 
             _replacements.add(new ReplacementParam<>("messageBody", String.class, "Message content, formatted as HTML", ContentType.HTML)
             {
+                @Override
                 public String getValue(Container c)
                 {
                     if (notificationBean == null)
@@ -1013,6 +1021,7 @@ public class AnnouncementManager
 
             _replacements.add(new ReplacementParam<>("messageBodyText", String.class, "Message content plain text", ContentType.Plain)
             {
+                @Override
                 public String getValue(Container c)
                 {
                     if (notificationBean == null)
@@ -1023,6 +1032,7 @@ public class AnnouncementManager
 
             _replacements.add(new ReplacementParam<>("messageSubject", String.class, "Message subject", ContentType.Plain)
             {
+                @Override
                 public String getValue(Container c)
                 {
                     if (notificationBean == null)
@@ -1033,6 +1043,7 @@ public class AnnouncementManager
 
             _replacements.add(new ReplacementParam<>("attachments", String.class, "Attachments for this message", ContentType.HTML)
             {
+                @Override
                 public String getValue(Container c)
                 {
                     return attachments;
@@ -1041,6 +1052,7 @@ public class AnnouncementManager
 
             _replacements.add(new ReplacementParam<>("reasonFooter", String.class, "Footer information explaining why user is receiving this message", ContentType.HTML)
             {
+                @Override
                 public String getValue(Container c)
                 {
                     return reasonForEmail;
@@ -1049,6 +1061,7 @@ public class AnnouncementManager
 
             _replacements.add(new ReplacementParam<>("emailPreferencesURL", String.class, "Link to allow users to configure their notification preferences", ContentType.Plain)
             {
+                @Override
                 public String getValue(Container c)
                 {
                     return emailPreferencesURL;
@@ -1074,6 +1087,7 @@ public class AnnouncementManager
             return super.renderTextBody(c);
         }
 
+        @Override
         public List<ReplacementParam> getValidReplacements()
         {
             return _replacements;
