+++ announcements/src/org/labkey/announcements/query/AnnouncementSchema.java
@@ -21,6 +21,7 @@ import org.labkey.announcements.model.AnnouncementManager;
 import org.labkey.api.announcements.CommSchema;
 import org.labkey.api.data.Container;
 import org.labkey.api.data.ContainerFilter;
+import org.labkey.api.data.CoreSchema;
 import org.labkey.api.data.SimpleFilter;
 import org.labkey.api.data.TableInfo;
 import org.labkey.api.module.Module;
@@ -50,12 +51,9 @@ public class AnnouncementSchema extends UserSchema
     public static final String FORUM_SUBSCRIPTION_TABLE_NAME = "ForumSubscription";
     public static final String ANNOUNCEMENT_SUBSCRIPTION_TABLE_NAME = "AnnouncementSubscription";
     public static final String EMAIL_OPTION_TABLE_NAME = "EmailOption";
-    public static final String EMAIL_FORMAT_TABLE_NAME = "EmailFormat";
     public static final String RSS_FEEDS_TABLE_NAME = "RSSFeeds";
     public static final String TOURS_TABLE_NAME = "Tours";
     public static final String THREADS_TABLE_NAME = "Threads";
-
-    public static final int PAGE_TYPE_ID = 0;
 
     private static final Set<String> TABLE_NAMES;
 
@@ -66,7 +64,6 @@ public class AnnouncementSchema extends UserSchema
         names.add(FORUM_SUBSCRIPTION_TABLE_NAME);
         names.add(ANNOUNCEMENT_SUBSCRIPTION_TABLE_NAME);
         names.add(EMAIL_OPTION_TABLE_NAME);
-        names.add(EMAIL_FORMAT_TABLE_NAME);
         names.add(RSS_FEEDS_TABLE_NAME);
         names.add(TOURS_TABLE_NAME);
         names.add(THREADS_TABLE_NAME);
@@ -85,6 +82,7 @@ public class AnnouncementSchema extends UserSchema
                 return !AnnouncementsController.getSettings(schema.getContainer()).isSecure();
             }
 
+            @Override
             public QuerySchema createSchema(DefaultSchema schema, Module module)
             {
                 return new AnnouncementSchema(schema.getUser(), schema.getContainer());
@@ -111,10 +109,6 @@ public class AnnouncementSchema extends UserSchema
         if (SPAM_TABLE_NAME.equalsIgnoreCase(name))
         {
             return createSpamTable(cf);
-        }
-        if (EMAIL_FORMAT_TABLE_NAME.equalsIgnoreCase(name))
-        {
-            return createEmailFormatTable(cf);
         }
         if (EMAIL_OPTION_TABLE_NAME.equalsIgnoreCase(name))
         {
@@ -143,18 +137,9 @@ public class AnnouncementSchema extends UserSchema
         return null;
     }
 
-    public TableInfo createEmailFormatTable(ContainerFilter cf)
-    {
-        FilteredTable result = new FilteredTable<>(CommSchema.getInstance().getTableInfoEmailFormats(), this, cf);
-        result.setName(EMAIL_FORMAT_TABLE_NAME);
-        result.wrapAllColumns(true);
-        result.setPublicSchemaName(getName());
-        return result;
-    }
-
     public TableInfo createEmailOptionTable(ContainerFilter cf)
     {
-        FilteredTable result = new FilteredTable<>(CommSchema.getInstance().getTableInfoEmailOptions(), this, cf);
+        FilteredTable result = new FilteredTable<>(CoreSchema.getInstance().getTableInfoEmailOptions(), this, cf);
         result.setName(EMAIL_OPTION_TABLE_NAME);
         result.addWrapColumn(result.getRealTable().getColumn("EmailOptionId"));
         result.addWrapColumn(result.getRealTable().getColumn("EmailOption"));
