+++ announcements/src/org/labkey/announcements/query/ForumSubscriptionTable.java
@@ -16,12 +16,12 @@
 package org.labkey.announcements.query;
 
 import org.jetbrains.annotations.NotNull;
-import org.labkey.api.announcements.CommSchema;
 import org.labkey.api.collections.CaseInsensitiveHashMap;
 import org.labkey.api.data.Container;
 import org.labkey.api.data.ContainerFilter;
 import org.labkey.api.data.ContainerForeignKey;
 import org.labkey.api.data.ContainerManager;
+import org.labkey.api.data.CoreSchema;
 import org.labkey.api.data.SimpleFilter;
 import org.labkey.api.data.Table;
 import org.labkey.api.data.TableInfo;
@@ -51,7 +51,7 @@ public class ForumSubscriptionTable extends AbstractSubscriptionTable
 {
     public ForumSubscriptionTable(AnnouncementSchema schema, ContainerFilter cf)
     {
-        super(CommSchema.getInstance().getTableInfoEmailPrefs(), schema, cf);
+        super(CoreSchema.getInstance().getTableInfoEmailPrefs(), schema, cf);
 
         var folderColumn = wrapColumn("Folder", getRealTable().getColumn("Container"));
         addColumn(folderColumn);
@@ -73,17 +73,6 @@ public class ForumSubscriptionTable extends AbstractSubscriptionTable
         });
         addColumn(emailOptionColumn);
 
-        var emailFormatColumn = wrapColumn("EmailFormat", getRealTable().getColumn("EmailFormatId"));
-        emailFormatColumn.setFk(new LookupForeignKey(cf, "EmailFormatId", null)
-        {
-            @Override
-            public TableInfo getLookupTableInfo()
-            {
-                return _userSchema.createEmailFormatTable(getLookupContainerFilter());
-            }
-        });
-        addColumn(emailFormatColumn);
-
         addCondition(getRealTable().getColumn("Type"), "messages");
 
         addWrapColumn(getRealTable().getColumn("SrcIdentifier"));
@@ -95,6 +84,7 @@ public class ForumSubscriptionTable extends AbstractSubscriptionTable
         return hasPermission(user, perm, getContainer());
     }
 
+    @Override
     public boolean hasPermission(UserPrincipal user, Class<? extends Permission> perm, Container container)
     {
         // Guests can't subscribe to anything, or edit anyone else's subscriptions, but they can read the table
@@ -223,7 +213,7 @@ public class ForumSubscriptionTable extends AbstractSubscriptionTable
         @Override
         protected Map<String, Object> getRow(User user, Container container, Map<String, Object> keys) throws InvalidKeyException
         {
-            return new TableSelector(ForumSubscriptionTable.this).getObject(getTargets(keys, user, container).createUserSchemaFilter(), Map.class);
+            return new TableSelector(ForumSubscriptionTable.this).getMap(getTargets(keys, user, container).createUserSchemaFilter());
         }
 
         @Override
@@ -240,7 +230,7 @@ public class ForumSubscriptionTable extends AbstractSubscriptionTable
 
                 Map<String, Object> insertMap = createDatabaseMap(user, row, targets);
 
-                Table.insert(user, CommSchema.getInstance().getTableInfoEmailPrefs(), insertMap);
+                Table.insert(user, CoreSchema.getInstance().getTableInfoEmailPrefs(), insertMap);
 
                 return getRow(user, container, row);
             }
@@ -248,7 +238,6 @@ public class ForumSubscriptionTable extends AbstractSubscriptionTable
             {
                 throw new QueryUpdateServiceException(e);
             }
-
         }
 
         /** Translate to the real database column names */
@@ -263,11 +252,6 @@ public class ForumSubscriptionTable extends AbstractSubscriptionTable
             {
                 insertMap.put("EmailOptionId", row.get("EmailOption"));
             }
-            if (row.containsKey("EmailFormat"))
-            {
-                insertMap.put("EmailFormatId", row.get("EmailFormat"));
-            }
-            insertMap.put("PageTypeId", AnnouncementSchema.PAGE_TYPE_ID);
             return insertMap;
         }
 
@@ -284,7 +268,7 @@ public class ForumSubscriptionTable extends AbstractSubscriptionTable
                 pks.put("Container", oldTargets.getContainer().getEntityId());
                 pks.put("Type", "messages");
                 pks.put("SrcIdentifier", oldTargets.getSrcIdentifier());
-                Table.update(user, CommSchema.getInstance().getTableInfoEmailPrefs(), createDatabaseMap(user, row, newTargets), pks);
+                Table.update(user, CoreSchema.getInstance().getTableInfoEmailPrefs(), createDatabaseMap(user, row, newTargets), pks);
             }
 
             return getRow(user, container, row);
@@ -294,7 +278,7 @@ public class ForumSubscriptionTable extends AbstractSubscriptionTable
         protected Map<String, Object> deleteRow(User user, Container container, Map<String, Object> oldRow) throws InvalidKeyException
         {
             SubscriptionTarget targets = getTargets(oldRow, user, container);
-            Table.delete(CommSchema.getInstance().getTableInfoEmailPrefs(), targets.createDbSchemaFilter());
+            Table.delete(CoreSchema.getInstance().getTableInfoEmailPrefs(), targets.createDbSchemaFilter());
 
             return oldRow;
         }
