+++ api/src/org/labkey/api/ApiModule.java
@@ -15,30 +15,68 @@
  */
 package org.labkey.api;
 
-import com.google.common.collect.ImmutableSet;
 import org.jetbrains.annotations.NotNull;
+import org.labkey.api.action.ApiXmlWriter;
+import org.labkey.api.admin.SubfolderWriter;
+import org.labkey.api.assay.ReplacedRunFilter;
 import org.labkey.api.attachments.AttachmentService;
 import org.labkey.api.attachments.LookAndFeelResourceType;
 import org.labkey.api.attachments.SecureDocumentType;
-import org.labkey.api.data.ExcelColumn;
-import org.labkey.api.data.SqlScanner;
-import org.labkey.api.data.WorkbookContainerType;
+import org.labkey.api.collections.ArrayListMap;
+import org.labkey.api.collections.CaseInsensitiveHashMap;
+import org.labkey.api.collections.CaseInsensitiveHashSet;
+import org.labkey.api.collections.CaseInsensitiveMapWrapper;
+import org.labkey.api.collections.CollectionUtils;
+import org.labkey.api.collections.MultiValuedMapCollectors;
+import org.labkey.api.collections.Sampler;
+import org.labkey.api.collections.SwapQueue;
+import org.labkey.api.data.*;
+import org.labkey.api.data.dialect.ParameterSubstitutionTest;
+import org.labkey.api.data.dialect.StandardDialectStringHandler;
+import org.labkey.api.dataiterator.CachingDataIterator;
 import org.labkey.api.dataiterator.DataIteratorUtil;
+import org.labkey.api.dataiterator.RemoveDuplicatesDataIterator;
+import org.labkey.api.dataiterator.ResultSetDataIterator;
+import org.labkey.api.dataiterator.SimpleTranslator;
+import org.labkey.api.dataiterator.StatementDataIterator;
+import org.labkey.api.exp.api.StorageProvisioner;
+import org.labkey.api.iterator.MarkableIterator;
 import org.labkey.api.jsp.LabKeyJspFactory;
+import org.labkey.api.markdown.MarkdownService;
 import org.labkey.api.module.CodeOnlyModule;
+import org.labkey.api.module.FolderTypeManager;
 import org.labkey.api.module.ModuleContext;
+import org.labkey.api.module.ModuleDependencySorter;
+import org.labkey.api.module.ModuleHtmlView;
+import org.labkey.api.query.AbstractQueryUpdateService;
+import org.labkey.api.query.AliasManager;
 import org.labkey.api.query.FieldKey;
 import org.labkey.api.query.SchemaKey;
+import org.labkey.api.reader.ExcelFactory;
+import org.labkey.api.reader.ExcelLoader;
+import org.labkey.api.reader.JSONDataLoader;
+import org.labkey.api.reader.MapLoader;
+import org.labkey.api.reader.TabLoader;
+import org.labkey.api.reports.model.ViewCategoryManager;
+import org.labkey.api.reports.report.RReport;
 import org.labkey.api.reports.report.ReportType;
 import org.labkey.api.security.ApiKeyManager;
 import org.labkey.api.security.ApiKeyManager.ApiKeyMaintenanceTask;
 import org.labkey.api.security.AuthenticationLogoType;
 import org.labkey.api.security.AuthenticationManager;
 import org.labkey.api.security.AvatarType;
+import org.labkey.api.security.Encryption;
+import org.labkey.api.security.GroupManager;
+import org.labkey.api.security.NestedGroupsTest;
+import org.labkey.api.security.PasswordExpiration;
+import org.labkey.api.security.SecurityManager;
+import org.labkey.api.security.ValidEmail;
 import org.labkey.api.settings.AppProps;
-import org.labkey.api.util.ContextListener;
-import org.labkey.api.util.Pair;
-import org.labkey.api.util.SystemMaintenance;
+import org.labkey.api.util.*;
+import org.labkey.api.util.emailTemplate.EmailTemplate;
+import org.labkey.api.view.ActionURL;
+import org.labkey.api.view.JspTemplate;
+import org.labkey.api.view.Portal;
 import org.labkey.api.view.WebPartFactory;
 
 import javax.servlet.jsp.JspFactory;
@@ -90,23 +128,126 @@ public class ApiModule extends CodeOnlyModule
     @Override
     public @NotNull Set<Class> getUnitTests()
     {
-        return ImmutableSet.of(
+        return Set.of(
+            Aggregate.TestCase.class,
+            ApiXmlWriter.TestCase.class,
+            ArrayListMap.TestCase.class,
+            BooleanFormat.TestCase.class,
+            BuilderObjectFactory.TestCase.class,
+            CachingDataIterator.ScrollTestCase.class,
+            CaseInsensitiveHashMap.TestCase.class,
+            CaseInsensitiveHashSet.TestCase.class,
+            CaseInsensitiveMapWrapper.TestCase.class,
+            ChecksumUtil.TestCase.class,
+            CollectionUtils.TestCase.class,
+            Compress.TestCase.class,
             Constants.TestCase.class,
+            ConvertHelper.TestCase.class,
+            DatabaseCache.TestCase.class,
             DataIteratorUtil.TestCase.class,
-            SqlScanner.TestCase.class,
+            DateUtil.TestCase.class,
+            DbScope.DialectTestCase.class,
+            EmailTemplate.TestCase.class,
+            ExcelFactory.ExcelFactoryTestCase.class,
+            ExcelLoader.ExcelLoaderTestCase.class,
+            ExtUtil.TestCase.class,
             FieldKey.TestCase.class,
+            FileType.TestCase.class,
+            FileUtil.TestCase.class,
+            HelpTopic.TestCase.class,
+            InlineInClauseGenerator.TestCase.class,
+            JSONDataLoader.HeaderMatchTest.class,
+            JSONDataLoader.MetadataTest.class,
+            JSONDataLoader.RowTest.class,
+            JsonTest.class,
+            MarkableIterator.TestCase.class,
+            MaterializedQueryHelper.TestCase.class,
+            MemTracker.TestCase.class,
+            ModuleDependencySorter.TestCase.class,
+            MultiValuedRenderContext.TestCase.class,
+            NumberUtilsLabKey.TestCase.class,
+            PageFlowUtil.TestCase.class,
+            Pair.TestCase.class,
+            PasswordExpiration.TestCase.class,
+            Path.TestCase.class,
+            RemoveDuplicatesDataIterator.DeDuplicateTestCase.class,
+            ReplacedRunFilter.TestCase.class,
+            ResultSetUtil.TestCase.class,
+            RReport.TestCase.class,
+            Sampler.TestCase.class,
             SchemaKey.TestCase.class,
-            Pair.TestCase.class
+            SimpleFilter.BetweenClauseTestCase.class,
+            SimpleFilter.FilterTestCase.class,
+            SimpleFilter.InClauseTestCase.class,
+            SqlScanner.TestCase.class,
+            StringExpressionFactory.TestCase.class,
+            StringUtilsLabKey.TestCase.class,
+            SubfolderWriter.TestCase.class,
+            SwapQueue.TestCase.class,
+            TabLoader.HeaderMatchTest.class,
+            TidyUtil.TestCase.class,
+            TSVMapWriter.Tests.class,
+            TSVWriter.TestCase.class,
+            ValidEmail.TestCase.class
         );
     }
 
     @Override
     public @NotNull Set<Class> getIntegrationTests()
     {
-        return ImmutableSet.of(
+        return Set.of(
+            AbstractQueryUpdateService.TestCase.class,
+            ActionURL.TestCase.class,
+            AliasManager.TestCase.class,
             ApiKeyManager.TestCase.class,
-            WorkbookContainerType.TestCase.class,
-            ExcelColumn.TestCase.class
+            AtomicDatabaseInteger.TestCase.class,
+            ContainerDisplayColumn.TestCase.class,
+            ContainerFilter.TestCase.class,
+            ContainerManager.TestCase.class,
+            DbSchema.CachingTestCase.class,
+            DbSchema.DDLMethodsTestCase.class,
+            DbSchema.SchemaCasingTestCase.class,
+            DbSchema.TableSelectTestCase.class,
+            DbSchema.TransactionTestCase.class,
+            DbScope.GroupConcatTestCase.class,
+            DbScope.TransactionTestCase.class,
+            DbSequenceManager.TestCase.class,
+            DomTestCase.class,
+            Encryption.TestCase.class,
+            ExcelColumn.TestCase.class,
+            ExceptionUtil.TestCase.class,
+            FolderTypeManager.TestCase.class,
+            GroupManager.TestCase.class,
+            JspTemplate.TestCase.class,
+            MapLoader.MapLoaderTestCase.class,
+            MarkdownService.TestCase.class,
+            MimeMap.TestCase.class,
+            ModuleHtmlView.TestCase.class,
+            MultiValuedMapCollectors.TestCase.class,
+            NestedGroupsTest.class,
+            ParameterSubstitutionTest.class,
+            Portal.TestCase.class,
+            PropertyManager.TestCase.class,
+            //RateLimiter.TestCase.class,
+            ResultSetDataIterator.TestCase.class,
+            ResultSetSelectorTestCase.class,
+            RowTrackingResultSetWrapper.TestCase.class,
+            SecurityManager.TestCase.class,
+            SimpleTranslator.TranslateTestCase.class,
+            SQLFragment.TestCase.class,
+            SqlSelectorTestCase.class,
+            StandardDialectStringHandler.TestCase.class,
+            StatementDataIterator.TestCase.class,
+            StatementUtils.TestCase.class,
+            StorageProvisioner.TestCase.class,
+            Table.DataIteratorTestCase.class,
+            Table.TestCase.class,
+            TableSelectorTestCase.class,
+            TabLoader.TabLoaderTestCase.class,
+            TempTableInClauseGenerator.TestCase.class,
+            URLHelper.TestCase.class,
+            ViewCategoryManager.TestCase.class,
+            WorkbookContainerType.TestCase.class
         );
     }
