+++ api/src/org/labkey/api/data/dialect/SqlDialectManager.java
@@ -80,10 +80,6 @@ public class SqlDialectManager
         for (SqlDialectFactory factory : FACTORIES)
             classes.addAll(factory.getJUnitTests());
 
-        // TODO: Why don't we register these directly in the module?
-        classes.add(ParameterSubstitutionTest.class);
-        classes.add(StandardDialectStringHandler.TestCase.class);
-
         return classes;
     }
