+++ api/src/org/labkey/api/security/GroupMembershipCache.java
@@ -23,6 +23,7 @@ import org.labkey.api.data.CoreSchema;
 import org.labkey.api.data.SQLFragment;
 import org.labkey.api.data.Selector;
 import org.labkey.api.data.SqlSelector;
+import org.labkey.api.security.UserManager.UserListener;
 
 import java.beans.PropertyChangeEvent;
 import java.util.Arrays;
@@ -46,7 +47,8 @@ public class GroupMembershipCache
 
     static
     {
-        UserManager.addUserListener(new GroupMembershipUserListener());
+        // Need to clear the cache before any other listener is called
+        UserManager.addUserListener(new GroupMembershipUserListener(), true);
     }
 
 
@@ -183,12 +185,14 @@ public class GroupMembershipCache
     }
 
 
-    public static class GroupMembershipUserListener implements UserManager.UserListener
+    public static class GroupMembershipUserListener implements UserListener
     {
+        @Override
         public void userAddedToSite(User user)
         {
         }
 
+        @Override
         public void userDeletedFromSite(User user)
         {
             // Blow away groups immediately after user is deleted, otherwise this user's groups, and therefore permissions, will remain active
@@ -196,12 +200,14 @@ public class GroupMembershipCache
             uncache(user);
         }
 
+        @Override
         public void userAccountDisabled(User user)
         {
             uncache(user);
 
         }
 
+        @Override
         public void userAccountEnabled(User user)
         {
             uncache(user);
