+++ api/src/org/labkey/api/security/UserManager.java
@@ -126,7 +126,16 @@ public class UserManager
     /** Adds a listener to be notified when user account actions happen */
     public static void addUserListener(UserListener listener)
     {
-        _listeners.add(listener);
+        addUserListener(listener, false);
+    }
+
+    /** Adds a listener with option to specify that it needs to be executed before the other listeners */
+    public static void addUserListener(UserListener listener, boolean meFirst)
+    {
+        if (meFirst)
+            _listeners.add(0, listener);
+        else
+            _listeners.add(listener);
     }
 
     private static List<UserListener> getListeners()
