+++ core/api-src/org/labkey/api/products/ProductRegistry.java
@@ -32,8 +32,6 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
-import java.util.stream.Collectors;
 
 public class ProductRegistry
 {
