+++ core/resources/schemas/core.xml
@@ -737,4 +737,23 @@
       <column columnName="EncryptedProperties"/>
     </columns>
   </table>
+  <table tableName="EmailOptions" tableDbType="TABLE">
+    <description>Contains one row per subscription type, such as all messages in a given forum, no email at all, etc and daily digest options</description>
+    <columns>
+      <column columnName="EmailOptionId"/>
+      <column columnName="EmailOption"/>
+      <column columnName="Type"/>
+    </columns>
+  </table>
+  <table tableName="EmailPrefs" tableDbType="TABLE">
+    <description>Contains one row per user-specific subscription</description>
+    <columns>
+      <column columnName="Container"/>
+      <column columnName="UserId"/>
+      <column columnName="Type"/>
+      <column columnName="EmailOptionId"/>
+      <column columnName="LastModifiedBy"/>
+      <column columnName="SrcIdentifier"/>
+    </columns>
+  </table>
 </tables>
