+++ core/src/org/labkey/core/admin/AdminController.java
@@ -87,12 +87,16 @@ import org.labkey.api.compliance.ComplianceService;
 import org.labkey.api.data.*;
 import org.labkey.api.data.Container.ContainerException;
 import org.labkey.api.data.queryprofiler.QueryProfiler;
+import org.labkey.api.data.queryprofiler.QueryProfiler.QueryStatTsvWriter;
 import org.labkey.api.exp.OntologyManager;
 import org.labkey.api.exp.api.StorageProvisioner;
 import org.labkey.api.exp.property.Lookup;
 import org.labkey.api.files.FileContentService;
+import org.labkey.api.message.settings.AbstractConfigTypeProvider.EmailConfigFormImpl;
 import org.labkey.api.message.settings.MessageConfigService;
 import org.labkey.api.message.settings.MessageConfigService.ConfigTypeProvider;
+import org.labkey.api.message.settings.MessageConfigService.NotificationOption;
+import org.labkey.api.message.settings.MessageConfigService.UserPreference;
 import org.labkey.api.miniprofiler.RequestInfo;
 import org.labkey.api.module.AllowedBeforeInitialUserIsSet;
 import org.labkey.api.module.AllowedDuringUpgrade;
@@ -109,7 +113,6 @@ import org.labkey.api.pipeline.PipelineStatusUrls;
 import org.labkey.api.pipeline.PipelineUrls;
 import org.labkey.api.pipeline.PipelineValidationException;
 import org.labkey.api.pipeline.view.SetupForm;
-import org.labkey.api.premium.PremiumService;
 import org.labkey.api.query.DefaultSchema;
 import org.labkey.api.query.FieldKey;
 import org.labkey.api.query.QuerySchema;
@@ -340,7 +343,8 @@ public class AdminController extends SpringActionController
             // Show Concepts tab only if the experiment module is enabled in this container
             return c.getActiveModules().contains(ModuleLoader.getInstance().getModule("Experiment"));
         }, AdminController.ConceptsAction.class);
-        addTab(TYPE.FolderManagement,"Notifications", "messages", NOT_ROOT, NotificationsAction.class);
+        // Show Notifications tab only if we have registered notification providers
+        addTab(TYPE.FolderManagement,"Notifications", "notifications", c->NOT_ROOT.test(c) && !MessageConfigService.get().getConfigTypes().isEmpty(), NotificationsAction.class);
         addTab(TYPE.FolderManagement,"Export", "export", NOT_ROOT, ExportFolderAction.class);
         addTab(TYPE.FolderManagement,"Import", "import", NOT_ROOT, ImportFolderAction.class);
         addTab(TYPE.FolderManagement,"Files", "files", FOLDERS_AND_PROJECTS, FileRootsAction.class);
@@ -2674,7 +2678,7 @@ public class AdminController extends SpringActionController
         @Override
         public void export(Object o, HttpServletResponse response, BindException errors) throws Exception
         {
-            try (QueryProfiler.QueryStatTsvWriter writer = new QueryProfiler.QueryStatTsvWriter())
+            try (QueryStatTsvWriter writer = new QueryStatTsvWriter())
             {
                 writer.setFilenamePrefix("SQL_Queries");
                 writer.write(response);
@@ -5642,9 +5646,9 @@ public class AdminController extends SpringActionController
             queryView.setButtonBarPosition(DataRegion.ButtonBarPosition.TOP);
 
             VBox defaultsView = new VBox(
-                    new HtmlView(
-                            "<div class=\"labkey-announcement-title\"><span>Default settings</span></div><div class=\"labkey-title-area-line\"></div>" +
-                                    "You can change this folder's default settings for email notifications here.")
+                new HtmlView(
+                    "<div class=\"labkey-announcement-title\"><span>Default settings</span></div><div class=\"labkey-title-area-line\"></div>" +
+                        "You can change this folder's default settings for email notifications here.")
             );
 
             PanelConfig config = new PanelConfig(getViewContext().getActionURL().clone(), key);
@@ -5654,17 +5658,16 @@ public class AdminController extends SpringActionController
             }
 
             return new VBox(
-                    new JspView<>("/org/labkey/core/admin/view/folderSettingsHeader.jsp", null, errors),
-                    defaultsView,
-                    new VBox(
-                            new HtmlView(
-                                    "<div class='labkey-announcement-title'><span>User settings</span></div><div class='labkey-title-area-line'></div>" +
-                                            "The list below contains all users with READ access to this folder who are able to receive notifications<br/>" +
-                                            "by email for message boards and file content events. A user's current message or file notification setting is<br/>" +
-                                            "visible in the appropriately named column.<br/><br/>" +
-                                            "To bulk edit individual settings: select one or more users, click the 'Update user settings' menu, and select the notification type."),
-                            queryView
-                    )
+                new JspView<>("/org/labkey/core/admin/view/folderSettingsHeader.jsp", null, errors),
+                defaultsView,
+                new VBox(
+                    new HtmlView(
+                        "<div class='labkey-announcement-title'><span>User settings</span></div><div class='labkey-title-area-line'></div>" +
+                            "The list below contains all users with read access to this folder who are able to receive notifications. Each user's current<br/>" +
+                            "notification setting is visible in the appropriately named column.<br/><br/>" +
+                            "To bulk edit individual settings: select one or more users, click the 'Update user settings' menu, and select the notification type."),
+                    queryView
+                )
             );
         }
 
@@ -5688,6 +5691,104 @@ public class AdminController extends SpringActionController
             }
             errors.reject(SpringActionController.ERROR_MSG, "Unable to find the selected config provider");
             return false;
+        }
+    }
+
+
+    public static class NotifyOptionsForm
+    {
+        private String _type;
+
+        public String getType()
+        {
+            return _type;
+        }
+
+        public void setType(String type)
+        {
+            _type = type;
+        }
+
+        public ConfigTypeProvider getProvider()
+        {
+            return MessageConfigService.get().getConfigType(getType());
+        }
+    }
+
+    /**
+     * Action to populate an Ext store with email notification options for admin settings
+     */
+    @RequiresPermission(AdminPermission.class)
+    public class GetEmailOptionsAction extends ReadOnlyApiAction<NotifyOptionsForm>
+    {
+        @Override
+        public ApiResponse execute(NotifyOptionsForm form, BindException errors)
+        {
+            ApiSimpleResponse resp = new ApiSimpleResponse();
+
+            ConfigTypeProvider provider = form.getProvider();
+            if (provider != null)
+            {
+                List<Map> options = new ArrayList<>();
+
+                // if the list of options is not for the folder default, add an option to use the folder default
+                if (getViewContext().get("isDefault") == null)
+                    options.add(PageFlowUtil.map("id", -1, "label", "Folder default"));
+
+                for (NotificationOption option : provider.getOptions())
+                {
+                    options.add(PageFlowUtil.map("id", option.getEmailOptionId(), "label", option.getEmailOption()));
+                }
+                resp.put("success", true);
+                if (!options.isEmpty())
+                    resp.put("options", options);
+            }
+            else
+                resp.put("success", false);
+
+            return resp;
+        }
+    }
+
+
+    @RequiresPermission(AdminPermission.class)
+    public class SetBulkEmailOptionsAction extends MutatingApiAction<EmailConfigFormImpl>
+    {
+        @Override
+        public ApiResponse execute(EmailConfigFormImpl form, BindException errors)
+        {
+            ApiSimpleResponse resp = new ApiSimpleResponse();
+            ConfigTypeProvider provider = form.getProvider();
+            String srcIdentifier = getContainer().getId();
+
+            Set<String> selections = DataRegionSelection.getSelected(getViewContext(), form.getDataRegionSelectionKey(), true);
+
+            if (!selections.isEmpty() && provider != null)
+            {
+                int newOption = form.getIndividualEmailOption();
+
+                for (String user : selections)
+                {
+                    User projectUser = UserManager.getUser(Integer.parseInt(user));
+                    UserPreference pref = provider.getPreference(getContainer(), projectUser, srcIdentifier);
+
+                    int currentEmailOption = pref != null ? pref.getEmailOptionId() : -1;
+
+                    //has this projectUser's option changed? if so, update
+                    //creating new record in EmailPrefs table if there isn't one, or deleting if set back to folder default
+                    if (currentEmailOption != newOption)
+                    {
+                        provider.savePreference(getUser(), getContainer(), projectUser, newOption, srcIdentifier);
+                    }
+                }
+                resp.put("success", true);
+            }
+            else
+            {
+                resp.put("success", false);
+                resp.put("message", "There were no users selected");
+            }
+            return resp;
         }
     }
