+++ filecontent/build.gradle
@@ -2,5 +2,4 @@ import org.labkey.gradle.util.BuildUtils
 
 dependencies{
     BuildUtils.addLabKeyDependency(project: project, config: "modules", depProjectPath: BuildUtils.getPlatformModuleProjectPath(project.gradle, "experiment"), depProjectConfig: "published", depExtension: "module")
-    BuildUtils.addLabKeyDependency(project: project, config: "modules", depProjectPath: BuildUtils.getPlatformModuleProjectPath(project.gradle, "announcements"), depProjectConfig: "published", depExtension: "module")
 }
