+++ internal/src/org/labkey/InternalModule.java
@@ -16,12 +16,18 @@
 package org.labkey;
 
 import org.jetbrains.annotations.NotNull;
+import org.labkey.api.data.TableViewFormTestCase;
+import org.labkey.api.exp.property.DomainTemplateGroup;
 import org.labkey.api.module.CodeOnlyModule;
 import org.labkey.api.module.ModuleContext;
+import org.labkey.api.script.RhinoService;
 import org.labkey.api.view.WebPartFactory;
+import org.labkey.api.webdav.ModuleStaticResolverImpl;
+import org.labkey.api.webdav.WebdavResolverImpl;
 
 import java.util.Collection;
 import java.util.Collections;
+import java.util.Set;
 
 /**
  * Created by susanh on 1/9/17.
@@ -43,5 +49,17 @@ public class InternalModule extends CodeOnlyModule
     @Override
     protected void doStartup(ModuleContext moduleContext)
     {
+    }
+
+    @Override
+    public @NotNull Set<Class> getIntegrationTests()
+    {
+        return Set.of(
+            DomainTemplateGroup.TestCase.class,
+            ModuleStaticResolverImpl.TestCase.class,
+            RhinoService.TestCase.class,
+            TableViewFormTestCase.class,
+            WebdavResolverImpl.TestCase.class
+        );
     }
 }
