+++ list/src/org/labkey/list/controllers/ListController.java
@@ -668,7 +668,7 @@ public class ListController extends SpringActionController
                 view.addView(new HtmlView(PageFlowUtil.textLink("show item history", getViewContext().cloneActionURL().addParameter("showHistory", "1"))));
             }
 
-            if (_list.getDiscussionSetting().isLinked() && LookAndFeelProperties.getInstance(getContainer()).isDiscussionEnabled())
+            if (_list.getDiscussionSetting().isLinked() && LookAndFeelProperties.getInstance(getContainer()).isDiscussionEnabled() && DiscussionService.get() != null)
             {
                 String entityId = item.getEntityId();
